requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.0
google-generativeai==0.3.0
PyPDF2==3.0.1
pdf2image==1.16.3
Pillow==10.1.0
python-pptx==0.6.21
PyMuPDF==1.23.8
qrcode==7.4.2
pywin32==305
python-dotenv==1.0.0
python-docx==0.8.11
gunicorn==21.2.0
